Create a basic command parser
The point is to have an easy way to define new commands that would match
how gulp could have some tasks executed before the main one.

A command do much, it's supposed to be a definition only. It orders a serie
of actions. If something more complex has to be done, it should be extracted
elsewhere in order to be easy to make the command parser a bit more savvy.