Revert 241 master (#263)
* Create a basic command parser

The point is to have an easy way to define new commands that would match
how gulp could have some tasks executed before the main one.

A command do much, it's supposed to be a definition only. It orders a serie
of actions. If something more complex has to be done, it should be extracted
elsewhere in order to be easy to make the command parser a bit more savvy.

* Create a basic CLI file.

The usage should be updated each time one of the task has been fully rewritten.

The CLI is still really ugly for now. `Chalk` could be a good choice in order
to add some fanciness.

* Update how command arguments are passed down to execution

* Rewrite config generator with new command system

* Fix node v0.10 compatibility

* Add promise system to handle command failures

* Execute all before/after scripts at the same time for performance reasons

* Fix promise propagation during command execution

* Enable internal command execution by using the prefix '_'

The reason I am adding this feature is because some commands need
other commands.

* Move gulp commands to `core/command` directory

Gulp dependency is not removed yet since but the commands are working

* Promisify commands in order to solve the order of asynchronous calls

* Rename `paths` to `config` and make it a parameter

This is supposed to allow people to fully use the command line.
It is not fully operational yet (especially for the checksum checks).

* Revert "WIP : Move Gulp tasks to custom CLI #164"