Notification board in the README file (#1017)
I am adding a notification board section near the top of the readme file to inform users of the repo what is currently going on with the latest build. I was quite confused myself when I ran backstop this morning to find out that Chrome was not running anymore. I had to go through closed pull requests and commits to find out what caused the issue, which is a bit tedious. I think this space could be useful.
P.S. I put in the last version of backstop that worked for me but if there is a later stable version please update @garris