Node 20: Docker, package.json dependency updates, README enhancements, and general bugfixes and enhancements for developer experience. (#1523)
* style(remote): adjusts spacing/quotes for eslint rules (#34)

* feat(backstopjs): updates dev/dependencies to latest possible versions (#34)

* build(backstopjs): adds engines, type (commonjs), and bump package.json version (#34)

* build: adds 'docker-load' script for local testing (#34)

this loads the custom docker image, built with buildx, into local system container namespace. necessary when updating dockerfile without the option to publish to docker registry.

* fix: removes obsolete `genConfig` npm script (#34)

`genConfig` is now `makeConfig` and is ultimately used in `backstop init`. There is no need to test this separately, as it's part of the
overall flow of `init`, which is now used in npm's `integration-test` script, and passing.

* fix: removes obsolete 'echo' npm script

This was originally part of the `gulp` workflow, back in the `slimer` and `casper` days. https://github.com/garris/BackstopJS/blob/b00ab5bf358fcb3b77724ac4ef10b30e86b5188d/core/command/echo.js

* test: adds success, caution, and failure messaging to unit, integration, smoke, and sanity tests

This provides a final outcome message for all npm scripts for internal project testing. With tests that `exit 1`, a caution message is used, as opposed to a full-blown "FAIL".

* build: specifies npm@>=8 in package engines

also renames "docker-load" script "load-docker" to align with other script naming conventions

* build: updates and refactors README.md

* build(docker): updates Dockerfile with node:20-bullseye

* test: adds playwright smoke tests

* test: echo fail message if smoke tests succeed

* feat(docker): includes playwright and deps installation in Dockerfile

* test(backstop): updates caution message with advice to check logs